requirements.txt.txt
Flask==3.0.0
Flask-Cors==4.0.0
requests==2.31.0
python-whois==0.8.0
urllib3==2.0.7